backend/app/requirements.txt
fastapi
uvicorn
python-multipart
requests
python-dotenv
boto3
openai
pydantic
pydantic-settings
passlib[bcrypt]
pyjwt
pyotp
sqlmodel
bcrypt==4.1.2
